pyuul_requirements.txt
cached-property==1.5.2
certifi==2016.2.28
cffi==1.10.0
cycler==0.11.0
dataclasses==0.8
decorator==4.1.2
h5py==3.1.0
ipykernel==4.6.1
ipython==6.1.0
ipython-genutils==0.2.0
jedi==0.10.2
jupyter-client==5.1.0
jupyter-core==4.3.0
kiwisolver==1.3.1
matplotlib==3.3.4
mock==4.0.3
nose==1.3.7
numpy==1.19.5
olefile==0.44
pandas==0.20.3
pexpect==4.2.1
pickleshare==0.7.4
Pillow==8.4.0
prompt-toolkit==1.0.15
ptyprocess==0.5.2
pycparser==2.18
Pygments==2.2.0
pyparsing==3.0.7
python-dateutil==2.6.1
pytz==2017.2
pyuul==0.3.1
PyYAML==6.0
pyzmq==16.0.2
scipy==0.19.1
SIDEKIT==1.2.3
simplegeneric==0.8.1
six==1.10.0
Theano==1.0.5
torch==1.10.2
torchvision==0.11.2
tornado==4.5.2
traitlets==4.3.2
typing-extensions==4.1.1
wcwidth==0.1.7
xlwt==1.3.0
